Splitting test_mixture
