+++ scikits/learn/mixture/tests/test_gmm.py
@@ -10,7 +10,6 @@ from scipy import stats
 from scikits.learn import mixture
 from scikits.learn.datasets.samples_generator import generate_random_spd_matrix
 from scikits.learn.mixture import GMM, DPGMM, VBGMM
-from scikits.learn.mixture.dpgmm import log_normalize
 
 
 rng = np.random.RandomState(0)
@@ -44,11 +43,6 @@ def test_normalize_3D():
     for axis in range(3):
         Anorm = mixture.normalize(A, axis)
         assert np.all(np.allclose(Anorm.sum(axis), 1.0))
-
-def test_log_normalize():
-    v = np.array([0.1, 0.8, 0.01, 0.09])
-    a = np.log(2*v)
-    assert np.allclose(v, log_normalize(a))
 
 
 def test_sample_gaussian():
@@ -279,62 +273,6 @@ class TestGMMWithFullCovars(unittest.TestCase, GMMTester):
     model = GMM
 
 
-class TestDPGMMWithSphericalCovars(unittest.TestCase, GMMTester):
-    cvtype = 'spherical'
-    model = DPGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestDPGMMWithDiagonalCovars(unittest.TestCase, GMMTester):
-    cvtype = 'diag'
-    model = DPGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestDPGMMWithTiedCovars(unittest.TestCase, GMMTester):
-    cvtype = 'tied'
-    model = DPGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestDPGMMWithFullCovars(unittest.TestCase, GMMTester):
-    cvtype = 'full'
-    model = DPGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-
-class TestVBGMMWithSphericalCovars(unittest.TestCase, GMMTester):
-    cvtype = 'spherical'
-    model = VBGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestVBGMMWithDiagonalCovars(unittest.TestCase, GMMTester):
-    cvtype = 'diag'
-    model = VBGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestVBGMMWithTiedCovars(unittest.TestCase, GMMTester):
-    cvtype = 'tied'
-    model = VBGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
-
-class TestVBGMMWithFullCovars(unittest.TestCase, GMMTester):
-    cvtype = 'full'
-    model = VBGMM
-
-    def score(self, g, train_obs):
-        return g.lower_bound()
 
 if __name__ == '__main__':
     nose.runmodule()
